Review comments of #3314, annotation based models. New test suite and fixed/worked on Mateusz' comments
