+++ config/src/test/java/org/axonframework/config/testsuite/student/commands/EnrollStudentToCourseCommand.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2010-2025. Axon Framework
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.axonframework.config.testsuite.student.commands;
+
+import org.axonframework.modelling.command.annotation.TargetModelIdentifier;
+
+public record EnrollStudentToCourseCommand(
+        @TargetModelIdentifier
+        String studentId,
+        String courseId
+) {
+
+}
